refactor(feed_builder): consolidate format builders under feed builder namespace
